Animes V2

+++ animes-renamer.sh
@@ -10,7 +10,7 @@ function get-mal-id () {
 jq ".[] | select( .tvdb_id == ${tvdb_id} )" -r $SCRIPT_FOLDER/tmp/pmm_anime_ids.json | jq .mal_id | sort -n | head -1
 }
 function get-mal-infos () {
-if [ ! -f $SCRIPT_FOLDER/data/$mal_id.json ]
+if [ ! -f $SCRIPT_FOLDER/data/$mal_id.json ] 										#check if exist
 then
 	sleep 0.5
 	curl "https://api.jikan.moe/v4/anime/$mal_id" > $SCRIPT_FOLDER/data/$mal_id.json 
@@ -24,7 +24,7 @@ function get-mal-rating () {
 jq .data.score -r $SCRIPT_FOLDER/data/$mal_id.json
 }
 function get-mal-poster () {
-if [ ! -f $SCRIPT_FOLDER/posters/$mal_id.jpg ]
+if [ ! -f $SCRIPT_FOLDER/posters/$mal_id.jpg ]										#check if exist
 then
 sleep 0.5
 	mal_poster_url=$(jq .data.images.jpg.large_image_url -r $SCRIPT_FOLDER/data/$mal_id.json)
@@ -40,23 +40,23 @@ jq ".[] | select( .mal_id == ${mal_id} )" -r $SCRIPT_FOLDER/tmp/pmm_anime_ids.js
 }
 
 # download pmm animes mapping and check if files and folder exist
-if [ ! -f $animes_titles ]
+if [ ! -f $animes_titles ]											#check if metadata files exist and echo first line
 then
         echo "metadata:" > $animes_titles
 fi
-if [ ! -d $SCRIPT_FOLDER/data ]
+if [ ! -d $SCRIPT_FOLDER/data ]											#check if exist and create folder for json data
 then
         mkdir $SCRIPT_FOLDER/data
 else
-	find $SCRIPT_FOLDER/data/* -mtime +2 -exec rm {} \;
+	find $SCRIPT_FOLDER/data/* -mtime +2 -exec rm {} \;							#delete json data if older than 2 days
 fi
-if [ ! -d $SCRIPT_FOLDER/posters ]
+if [ ! -d $SCRIPT_FOLDER/posters ]										#check if exist and create folder for posters
 then
         mkdir $SCRIPT_FOLDER/posters
 else
-	find $SCRIPT_FOLDER/posters/* -mtime +30 -exec rm {} \;
+	find $SCRIPT_FOLDER/posters/* -mtime +30 -exec rm {} \;							#delete posters if older than 30 days
 fi
-if [ ! -d $SCRIPT_FOLDER/ID ]
+if [ ! -d $SCRIPT_FOLDER/ID ]											#check if exist and create folder and file for ID
 then
 	mkdir $SCRIPT_FOLDER/ID
 	touch $SCRIPT_FOLDER/ID/animes.tsv
@@ -64,27 +64,27 @@ elif [ ! -f $SCRIPT_FOLDER/ID/animes.tsv ]
 then
 	touch $SCRIPT_FOLDER/ID/animes.tsv
 fi
-if [ ! -d $SCRIPT_FOLDER/tmp ]
+if [ ! -d $SCRIPT_FOLDER/tmp ]											#check if exist and create temp folder cleaned at the start of every run
 then
         mkdir $SCRIPT_FOLDER/tmp
 else
 	rm $SCRIPT_FOLDER/tmp/*
 fi
-curl "https://raw.githubusercontent.com/meisnate12/Plex-Meta-Manager-Anime-IDs/master/pmm_anime_ids.json" > $SCRIPT_FOLDER/tmp/pmm_anime_ids.json
+curl "https://raw.githubusercontent.com/meisnate12/Plex-Meta-Manager-Anime-IDs/master/pmm_anime_ids.json" > $SCRIPT_FOLDER/tmp/pmm_anime_ids.json		#download local copy of ID mapping
 
-# create pmm meta.log
+# Dummy run of PMM and move meta.log for creating tvdb_id and title_plex
 rm $PMM_FOLDER/config/temp-animes.cache
 $PMM_FOLDER/pmm-venv/bin/python3 $PMM_FOLDER/plex_meta_manager.py -r --config $PMM_FOLDER/config/temp-animes.yml
 mv $PMM_FOLDER/config/logs/meta.log $SCRIPT_FOLDER/tmp
 
-# create clean list-animes.csv (tvdb_id | title_plex) from meta.log
+# create clean list-animes.tsv (tvdb_id	title_plex) from meta.log
 line_start=$(grep -n "Mapping Animes Library" $SCRIPT_FOLDER/tmp/meta.log | cut -d : -f 1)
 line_end=$(grep -n -m1 "Animes Library Operations" $SCRIPT_FOLDER/tmp/meta.log | cut -d : -f 1)
 head -n $line_end $SCRIPT_FOLDER/tmp/meta.log | tail -n $(( $line_end - $line_start - 1 )) | head -n -5 > $SCRIPT_FOLDER/tmp/cleanlog-animes.txt
 awk -F"|" '{ OFS = "\t" } ; { gsub(/ /,"",$5) } ; { print substr($5,8),substr($7,2,length($7)-2) }' $SCRIPT_FOLDER/tmp/cleanlog-animes.txt > $SCRIPT_FOLDER/tmp/list-animes.tsv
 
-# create ID/animes.tsv ( tvdb_id | mal_id | title_mal | title_plex )
-while IFS=$'\t' read -r tvdb_id mal_id title_mal
+# create ID/animes.tsv from the clean list ( tvdb_id	mal_id	title_mal	title_plex )
+while IFS=$'\t' read -r tvdb_id mal_id title_mal								# First add the override animes to the ID file
 do
 	if ! awk -F"|" '{print $1}' $SCRIPT_FOLDER/ID/animes.tsv | grep "\<$tvdb_id\>"
 	then
@@ -94,12 +94,12 @@ do
 		echo "$(date +%H:%M:%S) - override found for : $title_mal / $title_plex" >> $LOG
 	fi
 done < $SCRIPT_FOLDER/override-ID-animes.tsv
-while IFS=$'\t' read -r tvdb_id title_plex
+while IFS=$'\t' read -r tvdb_id title_plex									# then get the other ID from the ID mapping and download json data
 do
 	if ! awk -F"\t" '{print $1}' $SCRIPT_FOLDER/ID/animes.tsv | grep "\<$tvdb_id\>"
 	then
 		mal_id=$(get-mal-id)
-		if [[ "$mal_title" == 'null' ]] || [[ "$mal_id" == 'null' ]] || [[ "${#mal_id}" == '0' ]]
+		if [[ "$mal_title" == 'null' ]] || [[ "$mal_id" == 'null' ]] || [[ "${#mal_id}" == '0' ]]	# Ignore anime with no tvdb to mal id conversion show in the error log you need to add them by hand in override
 		then
 			echo "$(date +%Y.%m.%d" - "%H:%M:%S) - invalid MAL ID for : tvdb : $tvdb_id / $title_plex" >> $ERROR_LOG
 		fi
@@ -110,30 +110,30 @@ do
 	fi
 done < $SCRIPT_FOLDER/tmp/list-animes.tsv
 
-#Create a currently ongoing list at $SCRIPT_FOLDER/data/ongoing.csv
-if [ ! -f $SCRIPT_FOLDER/data/ongoing.tsv ]
+#Create an ongoing list at $SCRIPT_FOLDER/data/ongoing.csv
+if [ ! -f $SCRIPT_FOLDER/data/ongoing.tsv ]		#check if already exist data folder is stored for 2 days 
 then
         ongoingpage=1
-        while [ $ongoingpage -lt 10 ];
+        while [ $ongoingpage -lt 10 ];			#get the airing list from jikan API max 9 pages (225 animes)
         do
                 curl "https://api.jikan.moe/v4/anime?status=airing&page=$ongoingpage&order_by=member&order=desc&genres_exclude=12&min_score=4" > $SCRIPT_FOLDER/tmp/ongoing-tmp.json
                 sleep 2
-                if grep "\"items\":{\"count\":0," $SCRIPT_FOLDER/tmp/ongoing-tmp.json
+                if grep "\"items\":{\"count\":0," $SCRIPT_FOLDER/tmp/ongoing-tmp.json			#stop if page is empty
                 then
                         break
                 fi
-		jq ".data[].mal_id" -r $SCRIPT_FOLDER/tmp/ongoing-tmp.json >> $SCRIPT_FOLDER/tmp/ongoing.tsv
+		jq ".data[].mal_id" -r $SCRIPT_FOLDER/tmp/ongoing-tmp.json >> $SCRIPT_FOLDER/tmp/ongoing.tsv		# store the mal ID of the ongoing show
                 ((ongoingpage++))
         done
         while read -r mal_id
         do
-                tvdb_id=$(get-tvdb-id)
-                if [[ "$tvdb_id" == 'null' ]] || [[ "${#tvdb_id}" == '0' ]]
+                tvdb_id=$(get-tvdb-id)											# convert the mal id to tvdb id (to get the main anime)
+                if [[ "$tvdb_id" == 'null' ]] || [[ "${#tvdb_id}" == '0' ]]						# Ignore anime with no mal to tvdb id conversion
                 then
-			echo "invalid TVDB ID for : MAL : $mal_id"
+			echo "Ongoing invalid TVDB ID for : MAL : $mal_id" >> $ERROR_LOG
 		else
-			if awk -F"|" '{print $1}' $SCRIPT_FOLDER/ID/animes.tsv | grep "\<$tvdb_id\>"
-			then
+				if awk -F"|" '{print $1}' $SCRIPT_FOLDER/ID/animes.tsv | grep "\<$tvdb_id\>"		# get the mal ID again but of the main anime
+				then
 				line=$(grep -n "\<$tvdb_id\>" $SCRIPT_FOLDER/ID/animes.tsv | cut -d : -f 1)
 				mal_id=$(sed -n "${line}p" $SCRIPT_FOLDER/ID/animes.tsv | awk -F"\t" '{print $2}')
 				title_mal=$(sed -n "${line}p" $SCRIPT_FOLDER/ID/animes.tsv | awk -F"\t" '{print $3}')
@@ -146,13 +146,13 @@ fi
 # write PMM metadata file from ID/animes.tsv and jikan API
 while IFS=$'\t' read -r tvdb_id mal_id title_mal title_plex
 do
-	if grep "\<$mal_id\>" $animes_titles
+	if grep "\<$mal_id\>" $animes_titles				# test if anime already in the metadata file and then replace some metadata
 	then
-		get-mal-infos
-		get-mal-poster
+		get-mal-infos						# check / download json data
+		get-mal-poster						# check / download poster
 		sorttitleline=$(grep -n "sort_title: \"$title_mal\"" $animes_titles | cut -d : -f 1)
 		ratingline=$((sorttitleline+1))
-		if sed -n "${ratingline}p" $animes_titles | grep "audience_rating:"
+		if sed -n "${ratingline}p" $animes_titles | grep "audience_rating:"	# Replace rating (audience)
 		then
 			sed -i "${ratingline}d" $animes_titles
 			mal_score=$(get-mal-rating)
@@ -160,7 +160,7 @@ do
 			echo "$(date +%H:%M:%S) - $title_mal updated score : $mal_score" >> $LOG
 		fi
 		tagsline=$((sorttitleline+2))
-		if sed -n "${tagsline}p" $animes_titles | grep "genre.sync:"
+		if sed -n "${tagsline}p" $animes_titles | grep "genre.sync:"		# Replace tags (genres, themes and demographics from MAL)
 		then
 			sed -i "${tagsline}d" $animes_titles
 			mal_tags=$(get-mal-tags)
@@ -168,7 +168,7 @@ do
 			echo "$(date +%Y.%m.%d" - "%H:%M:%S) - $title_mal updated tags : $mal_tags" >> $LOG
 		fi
 		labelline=$((sorttitleline+3))
-		if sed -n "${labelline}p" $animes_titles | grep "label"
+		if sed -n "${labelline}p" $animes_titles | grep "label"			# replace the Ongoing label according to MAL airing list
 		then
 			sed -i "${labelline}d" $animes_titles
 			if awk -F"\t" '{print $2}' $SCRIPT_FOLDER/data/ongoing.tsv | grep "\<$mal_id\>"
@@ -180,23 +180,23 @@ do
 				echo "$(date +%Y.%m.%d" - "%H:%M:%S) - $title_mal removed to Ongoing" >> $LOG
 			fi
 		fi
-	else
+	else												# New anime need to write all metadata
 		get-mal-infos
 		echo "  \"$title_mal\":" >> $animes_titles
-                echo "    alt_title: \"$title_plex\"" >> $animes_titles
+                echo "    alt_title: \"$title_plex\"" >> $animes_titles		
                 echo "    sort_title: \"$title_mal\"" >> $animes_titles
 		score_mal=$(get-mal-rating)
-                echo "    audience_rating: $score_mal" >> $animes_titles
+                echo "    audience_rating: $score_mal" >> $animes_titles				# rating (audience)
 		mal_tags=$(get-mal-tags)
-		echo "    genre.sync: anime,${mal_tags}"  >> $animes_titles
-		if awk -F"\t" '{print $2}' $SCRIPT_FOLDER/data/ongoing.tsv | grep "\<$mal_id\>"
+		echo "    genre.sync: anime,${mal_tags}"  >> $animes_titles				# tags (genres, themes and demographics from MAL)
+		if awk -F"\t" '{print $2}' $SCRIPT_FOLDER/data/ongoing.tsv | grep "\<$mal_id\>"		# Ongoing label according to MAL airing list
 		then
 			echo "    label: Ongoing" >> $animes_titles
 		else
 			echo "    label.remove: Ongoing" >> $animes_titles
 		fi
-		get-mal-poster
-		echo "    file_poster: $SCRIPT_FOLDER/posters/${mal_id}.jpg" >> $animes_titles
+		get-mal-poster										# check / download poster
+		echo "    file_poster: $SCRIPT_FOLDER/posters/${mal_id}.jpg" >> $animes_titles		# add poster 
 		echo "#   mal_id: $mal_id" >> $animes_titles
 		echo "$(date +%H:%M:%S) - added to metadata : $title_mal / $title_plex / score : $score_mal / tags / poster" >> $LOG
 	fi
